[FIX] 임포트 오류 해결

+++ src/main/java/kr/hs/entrydsm/main/service/like/LikeService.java
@@ -4,6 +4,9 @@ import kr.hs.entrydsm.main.payload.response.PageResponse;
 import org.springframework.data.domain.Pageable;
 
 public interface LikeService {
-    void createLike(Integer postId);
+
+    void createLike(long postId);
+
     PageResponse bestPostList(Pageable page);
+
 }
